py_project2/requirements.txt
aniso8601==7.0.0
attrs==24.2.0
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==2.1.3
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
pip==21.2.3
promise==2.3
psycopg2-binary==2.9.10
python-dotenv==0.19.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
Rx==1.6.3
setuptools==57.4.0
six==1.17.0
SQLAlchemy==1.4.46
typing_extensions==4.12.2
Werkzeug==2.2.3

